% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_likelihood.R
\name{branching_point_process_likelihood}
\alias{branching_point_process_likelihood}
\title{Branching Point Process Likelihood}
\usage{
branching_point_process_likelihood(
  t,
  beta,
  a,
  nu,
  xi,
  omega = NULL,
  alpha = NULL,
  log = TRUE,
  perform_checks = TRUE
)
}
\arguments{
\item{t}{A real-valued vector of length N. The times of the point process.}

\item{beta}{An integer vector of length N. The branching structure of the
point process.}

\item{a}{A real valued scalar. The upper limit of the observation interval.}

\item{nu}{A positive real-valued vector of length N. The fitness associated
with each point of the point process.}

\item{xi}{A positive real-valued vector of length N. The memory decay
associated with each point of the point process.}

\item{omega}{A positive, increasing real-valued vector of length K.
Frequencies defining the sinusoidal basis functions for exogeneous
excitation.}

\item{alpha}{A real-valued vector of length 2K. Coefficients of the
sinusoidal basis function for exogenous excitation.}

\item{log}{A logical scalar. If `TRUE` then the log-likelihood is returned.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
The (log)-likelihood for the branching point process.
}
\description{
Computes the branching point process likelihood for the class of branching
processes described in Meagher & Friel.
}
