% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rps.R
\name{compute_pwm_crps}
\alias{compute_pwm_crps}
\title{Compute the PWM form of the CPRS}
\usage{
compute_pwm_crps(obs, pred, perform_checks = TRUE)
}
\arguments{
\item{obs}{A real-valued scalar. The observed value.}

\item{pred}{A real-valued vector. A set of forecasts.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A positive scalar value. Smaller values correspond to better probabilistic forecasts.
}
\description{
Computes the Probability Weighted Moment form of the Continuous Ranked Probability
Score, This is an unbiased estimate for the CRPS.
}
