% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rps.R
\name{compute_rps}
\alias{compute_rps}
\title{Compute the ranked probability score}
\usage{
compute_rps(
  obs,
  pred_prob,
  n_outcomes = max(length(pred_prob), obs),
  perform_checks = TRUE
)
}
\arguments{
\item{obs}{A positive integer scalar. The observed ordinal variable.}

\item{pred_prob}{A vector of probabilities. The forecast probability for each
possible outcome.}

\item{n_outcomes}{A positive integer scalar. The total number of possible
outcomes. The RPS is scaled according to `n_outcomes` and so specifying
this quantity explicitly allows model comparison without increasing the
computational overhead}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A scalar value on the unit interval. Smaller values correspond to
  better probabilistic forecasts.
}
\description{
The ranked probability score (RPS) is a proper scoring rule for evaluating
the quality of probabilistic forecasts for ordinal discrete variables.
}
