% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_models.R
\name{fit_branching_point_process}
\alias{fit_branching_point_process}
\title{Fit a branching point process model}
\usage{
fit_branching_point_process(
  t,
  branching_structure,
  observation_interval,
  point_type = rep(1, length(t)),
  K = 0,
  omega = numeric(0),
  is_hetero = rep(FALSE, length(unique(point_type))),
  shape_mu = 4,
  rate_mu = 8,
  shape_eta = 1,
  rate_eta = 1,
  perform_checks = TRUE,
  ...
)
}
\arguments{
\item{t}{A vector of N positive real values. The time associated with each point ordered by cluster.}

\item{branching_structure}{A vector of N non-negative integers. The underlying branching structure of the point process.
Note that this branching structure is defined within each cluster rather for the overall point process.}

\item{observation_interval}{A positive scalar or vector of length S. The interval over which each point process is observed. Defined either in terms of the end of observation
interval for each of the S clusters, or as a scalar defining the observation interval for each cluster from the initial immigrant point.}

\item{point_type}{An integer vector of length N. The pre-specified type of each point.}

\item{K}{A non-negative integer. The number of sinusoidal components in the
relative activity function,}

\item{omega}{A real valued vector of length K. The angular frequency of each sinusoidal
component.}

\item{is_hetero}{A logical vector of length M. Do points of each type have heterogeneous offspring
processes.}

\item{shape_mu}{A positive real-valued scalar. The shape of the Gamma prior on the reproduction number.}

\item{rate_mu}{A positive real-valued scalar. The rate of the Gamma prior on the reproduction number.}

\item{shape_eta}{A positive real-valued scalar. The shape of the Gamma prior on the memory decay rate.}

\item{rate_eta}{A positive real-valued scalar. The rate of the Gamma prior on the memory decay rate.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}

\item{...}{Additional arguments for`rstan::sampling`.}
}
\value{
A `stanfit` object. Samples from the posterior distribution over model parameters.
}
\description{
Fits the Branching point process models described in Meagher & Friel to data.
}
