% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_models.R
\name{fit_gpmcp}
\alias{fit_gpmcp}
\title{Fit a Gamma-Poisson-mixture cluster process (GPMCP) model}
\usage{
fit_gpmcp(
  t,
  branching_structure,
  type = rep(1, length(t)),
  a = floor(min(t)),
  b = ceiling(max(t)),
  is_hetero = FALSE,
  K = 0,
  f = numeric(0),
  sigma_R = 1,
  sigma_phi = 0.25,
  sigma_eta = 2,
  perform_checks = TRUE,
  pars = c("R", "phi", "eta", "beta", "A", "phase"),
  ...
)
}
\arguments{
\item{t}{An ordered vector of real-values. The time of each distinct event.}

\item{branching_structure}{A vector of non-negative integers. Immigrant
events identified by 0 while offspring link to the index of their parent.}

\item{type}{A vector of positive integers. Assigns each event to one of a set
number of specified types.}

\item{a}{A real-valued scalar. The start of the observation interval.}

\item{b}{A real-valued scalar. The end of the observation interval.}

\item{is_hetero}{Logical. Does the model allow for heterogeneous offspring
processes?}

\item{K}{A non-negative integer. The number of sinusoidal components in the
relative activity function,}

\item{f}{A real valued vector of length K. The frequency of each sinusoidal
component.}

\item{sigma_R}{A positive real-valued scalar. The standard deviation
parameter for the log-normal prior on R.}

\item{sigma_phi}{A positive real-valued scalar. The standard deviation
parameter for the log-normal prior on R.}

\item{sigma_eta}{A positive real-valued scalar. The standard deviation
parameter for the log-normal prior on R.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}

\item{pars}{The parameters returned by the sampling algorithm.}

\item{...}{Additional arguments for`rstan::sampling`.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Sample from the posterior of the TGPMCP model given event times and the
underlying branching structure.
}
