% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_models.R
\name{fit_periodic_point_process}
\alias{fit_periodic_point_process}
\title{Fit a periodic point process to data}
\usage{
fit_periodic_point_process(
  t,
  immigrant_observation_interval,
  K = 0,
  omega = numeric(0),
  perform_checks = TRUE,
  ...
)
}
\arguments{
\item{t}{A vector of N positive real values. The time associated with each point.}

\item{immigrant_observation_interval}{A real valued scalar. Defines the interval over which points arrive.}

\item{K}{A non-negative integer. The number of sinusoidal components in the
relative activity function,}

\item{omega}{A real valued vector of length K. The angular frequency of each sinusoidal
component.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}

\item{...}{Additional arguments for`rstan::sampling`.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Fits a Poisson process with periodic intensity to data.
}
