% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_TiDe-DSH.R
\name{fit_TiDeDSH}
\alias{fit_TiDeDSH}
\title{Fit a Time-Dependant Doubly-Stochastic Hawkes (TiDe-DSH) process model}
\usage{
fit_TiDeDSH(
  t,
  branching_structure,
  a = floor(min(t)),
  b = ceiling(max(t)),
  is_ds = FALSE,
  K = 0,
  f = numeric(0),
  perform_checks = TRUE,
  pars = c("mu", "R", "alpha", "eta", "beta", "A", "phase"),
  ...
)
}
\arguments{
\item{t}{An ordered vector of real-values. The time of each distinct event.}

\item{branching_structure}{A vector of non-negative integers. Immigrant events identified by 0 while
offspring link to the index of their parent.}

\item{a}{A real-valued scalar. The start of the observation interval.}

\item{b}{A real-valued scalar. The end of the observation interval.}

\item{is_ds}{Logical. Is the model doubly stochastic?}

\item{K}{A non-negative integer. The number of sinusoidal components in the relative activity function,}

\item{f}{A real valued vector of length K. The frequency of each sinusoidal component.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}

\item{pars}{The parameters return by the sampling algorithm.}

\item{...}{Additional arguments for`rstan::sampling`.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Sample from the posterior of the TiDe-DSH model given event times and the
underlying branching structure.
}
