% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_distributions.R
\name{gamma_posterior_individual_reproduction_number}
\alias{gamma_posterior_individual_reproduction_number}
\title{Individual reproduction number posterior}
\usage{
gamma_posterior_individual_reproduction_number(
  t,
  observed_offspring,
  observation_horizon,
  reproduction_number,
  dispersion_parameter,
  gi_exp_decay_rate,
  sinusoid_coefficients = NULL,
  sinusoid_frequencies = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{t}{A real-valued vector or scalar. The time points at which the event
arrived.}

\item{observed_offspring}{A non-negative integer vector or scalar. The number
of offsring from the observed event.}

\item{observation_horizon}{A real valued vector or scalar greater than
\eqn{t}. The time up to which offspring are observed,}

\item{reproduction_number}{A positive real valued scalar. The expected
number of offspring from each event.}

\item{dispersion_parameter}{A positive real valued scalar. The
dispersion parameter for offspring from each event.}

\item{gi_exp_decay_rate}{A positive real valued scalar. The exponential rate
of decay in the density function of the generation interval distribution,}

\item{sinusoid_coefficients}{A real-valued vector of length \eqn{2 * K}.
Denoted \eqn{\beta} above.}

\item{sinusoid_frequencies}{A positive real-valued vector of length \eqn{K}.
Angular frequencies of the sinusoidal basis. Denoted \eqn{\omega} above.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A named list of length 2 containing shape and rate parameters for the
  Gamma posterior distribution on individual reproduction numbers.
}
\description{
Parameterise the Gamma distributed posterior on individual reproduction
numbers.
}
