% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_likelihood.R
\name{heterogeneous_tide_mixture_bpp_likelihood}
\alias{heterogeneous_tide_mixture_bpp_likelihood}
\title{Heterogeneous Time-dependant Mixture Branching Process Likelihood}
\usage{
heterogeneous_tide_mixture_bpp_likelihood(
  t,
  beta,
  a,
  b = a,
  mu_immigrant,
  eta_immigrant,
  psi_immigrant,
  mu_offspring = mu_immigrant,
  eta_offspring = eta_immigrant,
  psi_offspring = psi_immigrant,
  omega = NULL,
  alpha = NULL,
  log = TRUE,
  n_samples = 100,
  seed = NULL,
  perform_checks = TRUE
)
}
\arguments{
\item{t}{A real-valued vector of length N. The times of the point process.}

\item{beta}{An integer vector of length N. The branching structure of the
point process.}

\item{a}{A real-valued scalar. The upper limit of the observation interval for estimating the latent fitness of each point.}

\item{b}{A real-valued scalar. The upper limit of the interval over which the likelihood is estimated.}

\item{mu_immigrant}{A positive real-valued scalar. The reproduction number
for immigrant offspring processes.}

\item{eta_immigrant}{A positive real-valued scalar. The memory decay rate for
immigrant offspring processes.}

\item{psi_immigrant}{A positive real valued scalar. The dispersion parameter for immigrant point fitness.}

\item{mu_offspring}{A positive real-valued scalar. The reproduction number
for offspring offspring processes.}

\item{eta_offspring}{A positive real-valued scalar. The memory decay rate for
offspring offspring processes.}

\item{psi_offspring}{A positive real valued scalar. The dispersion parameter for offspring point fitness.}

\item{omega}{A positive, increasing real-valued vector of length K.
Frequencies defining the sinusoidal basis functions for exogeneous
excitation.}

\item{alpha}{A real-valued vector of length 2K. Coefficients of the
sinusoidal basis function for exogenous excitation.}

\item{log}{A logical scalar. If `TRUE` then the log-likelihood is returned.}

\item{n_samples}{A positive integer. The number of latent fitness samples drawn to produce the Monte Carlo estimator.}

\item{seed}{The random seed for sampling fitness variables,}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A scalar value. A monte carlo estimator of the likelihood for the branching point process.
}
\description{
Computes the branching point process likelihood for the heterogeneous time-dependant
mixture branching process described in Meagher & Friel, whereby the offspring
processes for immigrants and offspring differ, follow a circadian rhythm,
and have heterogeneous fitness..
}
\seealso{
tide_mixture_bpp_likelihood homogeneous_bpp_likelihood mixture_bpp_likelihood
}
