% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_likelihood.R
\name{homogeneous_bpp_likelihood}
\alias{homogeneous_bpp_likelihood}
\title{Homogeneous Branching Point Process Likelihood}
\usage{
homogeneous_bpp_likelihood(
  t,
  beta,
  a,
  mu,
  eta,
  log = TRUE,
  perform_checks = TRUE
)
}
\arguments{
\item{t}{A real-valued vector of length N. The times of the point process.}

\item{beta}{An integer vector of length N. The branching structure of the
point process.}

\item{a}{A real valued scalar. The upper limit of the observation interval.}

\item{mu}{A positive real-valued scalar. The reproduction number for each offspring process.}

\item{eta}{A positive real-valued scalar. The memory decay rate for each offspring process.}

\item{log}{A logical scalar. If `TRUE` then the log-likelihood is returned.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A scalar value. The likelihood for the branching point process.
}
\description{
Computes the branching point process likelihood for the class of branching
processes described in Meagher & Friel.
}
\examples{
# Set a circadian rhythm
day <- 24
f <- 1 / c(day, day / 2)
omega <- 2 * pi * f
alpha <- c(-0.17, -0.59, -0.25, 0.34)
# Simulate data
set.seed(1010)
sim_cluster <- simulate_gpm_cluster_process(
  t_seed = 8, branching_structure_seed = 0,
  observation_horizon = 8 + 0, simulation_horizon = 8 + 48,
  immigrant_reproduction_number = 4,
  immigrant_gi_exp_decay_rate = 0.2,
  immigrant_dispersion_parameter = Inf,
  offspring_reproduction_number = 0.66,
  offspring_gi_exp_decay_rate = 0.33,
  offspring_dispersion_parameter = 0.1,
  sinusoid_coefficients = alpha,
  sinusoid_frequencies = omega
)
# Propose some parameters
m <- 0.66
r <- 0.3
# Evaluate the likelihood
homogeneous_bpp_likelihood(
  t = sim_cluster$t, beta = sim_cluster$parent_id, a = 8 + 48,
  mu = m, eta = r
)
}
