% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_structure.R
\name{identify_clusters}
\alias{identify_clusters}
\title{Identify and label clusters}
\usage{
identify_clusters(branching_structure)
}
\arguments{
\item{branching_structure}{The vector of non-negative integers that define
the branching structure.}
}
\value{
A vector of positive integers labelling members of each distinct
  cluster
}
\description{
Identify and label separate clusters identified from a branching structure
where \eqn{e_j = i} if the event at index \eqn{j} is the offspring of the
event at index \eqn{i} and \eqn{e_j = 0} otherwise. Note that each cluster is
centred on an immigrant event,
}
