% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_likelihood.R
\name{mixture_bpp_likelihood}
\alias{mixture_bpp_likelihood}
\title{Mixture Branching Process Likelihood}
\usage{
mixture_bpp_likelihood(
  t,
  beta,
  a,
  mu_immigrant,
  eta_immigrant,
  mu_offspring = mu_immigrant,
  eta_offspring = eta_immigrant,
  log = TRUE,
  perform_checks = TRUE
)
}
\arguments{
\item{t}{A real-valued vector of length N. The times of the point process.}

\item{beta}{An integer vector of length N. The branching structure of the
point process.}

\item{a}{A real valued scalar. The upper limit of the observation interval.}

\item{mu_immigrant}{A positive real-valued scalar. The reproduction number
for immigrant offspring processes.}

\item{eta_immigrant}{A positive real-valued scalar. The memory decay rate for
immigrant offspring processes.}

\item{mu_offspring}{A positive real-valued scalar. The reproduction number
for offspring offspring processes.}

\item{eta_offspring}{A positive real-valued scalar. The memory decay rate for
offspring offspring processes.}

\item{log}{A logical scalar. If `TRUE` then the log-likelihood is returned.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A scalar value. The likelihood for the branching point process.
}
\description{
Computes the branching point process likelihood for the mixture branching
process described in Meagher & Friel, whereby the offspring processes for
immigrants and offspring differ, but are otherwise homogeneous.
}
\seealso{
homogeneous_bpp_likelihood tide_mixture_bpp_likelihood
}
