% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineMessageboardActivity-package.R
\docType{package}
\name{onlineMessageboardActivity-package}
\alias{onlineMessageboardActivity-package}
\alias{onlineMessageboardActivity}
\title{The 'onlineMessageboardActivity' package.}
\description{
Poisson cluster processes, of which the Hawkes process is a
  special case, are useful family of models for online messageboard activity.
  This package includes algorithms for data manipulation, simulation and
  inference.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R
  package version 2.19.3. https://mc-stan.org
}
