% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_likelihood.R
\name{periodic_poisson_point_process_likelihood}
\alias{periodic_poisson_point_process_likelihood}
\title{Liklihood function for a Poisson process}
\usage{
periodic_poisson_point_process_likelihood(
  t,
  a,
  lambda_0,
  omega = NULL,
  alpha = NULL,
  log = TRUE,
  perform_checks = TRUE
)
}
\arguments{
\item{t}{A vector of N positive real values. The time associated with each point.}

\item{a}{A real valued scalar. Defines the interval over which points arrive.}

\item{lambda_0}{The average intensity of arrivals.}

\item{omega}{The frequencies of the sinusoidal basis functions for the periodic function.}

\item{alpha}{The coefficients on each sinusoidal basis function}

\item{log}{Logical. Report the log likelihood?}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A real value. The liklihood for the specified process.
}
\description{
Computes the likelihood for a Poisson process with a periodic intensity function
}
