% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_clusters.R
\name{simulate_ds_hawkes_cluster}
\alias{simulate_ds_hawkes_cluster}
\title{Simulate Doubly-Stochastic Hawkes process cluster}
\usage{
simulate_ds_hawkes_cluster(
  observed_cluster_df,
  reproduction_number,
  dispersion_parameter,
  gi_exp_decay_rate,
  observation_horizon,
  simulation_horizon,
  cluster_id = 0,
  perform_checks = TRUE
)
}
\arguments{
\item{observed_cluster_df}{A data frame containing the branching structure
and timestamps for events in the observation window. The observed events
seeding the cluster..}

\item{reproduction_number}{A positive real valued scalar. The expected number
of offspring from each event.}

\item{dispersion_parameter}{A positive real valued scalar. The dispersion
parameter for offspring from each event, denoted \eqn{k}.}

\item{gi_exp_decay_rate}{A positive real valued scalar. The exponential rate
of decay in the density function of the generation interval distribution,}

\item{observation_horizon}{A real valued scalar. The time up to which events
in the cluster are observed.}

\item{simulation_horizon}{A real valued scalar. The time up to which events
in the cluster are simulated.}

\item{cluster_id}{A scalar.An arbitrary cluster label.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A data frame with 4 variables: timestamps \eqn{t}, labels \eqn{id};
  parent labels \eqn{parent_id}; cluster label \eqn{cluster_id}.
}
\description{
Simulate a cluster generated by the doubly stochastic Hawkes process model up
to a specified horizon given the cluster over a specified observation window.
In this case, the offspring intensity is defined as \deqn{\nu_i \psi \exp (-
\psi ( t - t_i))} for \eqn{t > t_i} where \eqn{\nu_i} is the individual
reproduction number such that \eqn{E (\nu_i) = R} and \eqn{Var(\nu_i) = R +
R^2 / k}, and \eqn{\psi} is the exponential decay rate for the generation
interval density.
}
