% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gpmcp.R
\name{simulate_gpm_cluster_process}
\alias{simulate_gpm_cluster_process}
\title{Simulate a Gamma-Poisson Mixture Cluster Process}
\usage{
simulate_gpm_cluster_process(
  t_seed,
  branching_structure_seed,
  observation_horizon,
  simulation_horizon,
  immigrant_reproduction_number,
  immigrant_gi_exp_decay_rate,
  immigrant_dispersion_parameter = Inf,
  offspring_reproduction_number = immigrant_reproduction_number,
  offspring_gi_exp_decay_rate = immigrant_gi_exp_decay_rate,
  offspring_dispersion_parameter = immigrant_dispersion_parameter,
  sinusoid_coefficients = NULL,
  sinusoid_frequencies = NULL,
  dominating_scalar = ifelse(is.null(sinusoid_frequencies), 1, 2),
  cluster_id = NULL,
  refactor_output = TRUE,
  perform_checks = TRUE
)
}
\arguments{
\item{t_seed}{A real-valued vector. The time of the seed events.}

\item{branching_structure_seed}{A vector of non-negative integer values. The
branching structure of seed events.}

\item{observation_horizon}{A real valued scalar. The time up to which events
in the cluster are observed.}

\item{simulation_horizon}{A real valued scalar. The time up to which events
in the cluster are simulated.}

\item{immigrant_reproduction_number}{A positive real valued scalar. The
expected number of offspring from the immigrant event.}

\item{immigrant_gi_exp_decay_rate}{A positive real valued scalar. The
exponential rate of decay in the density function of the generation
interval distribution for the immigrant offspring process.}

\item{immigrant_dispersion_parameter}{A positive real valued scalar. The
dispersion parameter for offspring from the immigrant event. Offspring from
the immigrant event are Poisson distributed when this parameter is
Infinite.}

\item{offspring_reproduction_number}{A positive real valued scalar. The
expected number of offspring from offspring events.}

\item{offspring_gi_exp_decay_rate}{A positive real valued scalar. The
exponential rate of decay in the density function of the generation
interval distribution for offspring offspring process.}

\item{offspring_dispersion_parameter}{A positive real valued scalar. The
dispersion parameter for offspring from offspring events. Offspring from
the offspring events are Poisson distributed when this parameter is
Infinite.}

\item{sinusoid_coefficients}{A real-valued vector of length \eqn{2 * K}.
\eqn{NULL} default value indicates no absolute-time dependance in the
offspring process.}

\item{sinusoid_frequencies}{A positive real-valued vector of length \eqn{K}.
¿\eqn{NULL} default value indicates no absolute-time dependance in the
offspring process. Angular frequencies of the sinusoidal basis.}

\item{dominating_scalar}{A positive real valued scalar greater than the
absolute-time modulating function for all \eqn{t}. Allows Lewis thinning
for simulating offspring processes.}

\item{cluster_id}{A scalar.An arbitrary cluster label.}

\item{refactor_output}{Logical. Should the output branching structure be
refactored to a more natural representation. This refactoring is
computationally expensive and should be set to `FALSE` when it is not
required.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A data frame summarising the simulated Gamma-Poisson mixture cluster
  process.
}
\description{
Simulate a cluster from the family of generative modes described in Meagher &
Friel.
}
