% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gpmcp.R
\name{simulate_poisson_offspring_process}
\alias{simulate_poisson_offspring_process}
\title{Simulate a Poisson offspring process}
\usage{
simulate_poisson_offspring_process(
  t_seed,
  sampling_interval,
  reproduction_number,
  gi_exp_decay_rate,
  sinusoid_coefficients = NULL,
  sinusoid_frequencies = NULL,
  dominating_scalar = ifelse(is.null(sinusoid_frequencies), 1, 2),
  perform_checks = TRUE
)
}
\arguments{
\item{t_seed}{A real-valued scalar. The time of the seed event.}

\item{sampling_interval}{An ordered vector of length 2 with real values. The
interval over which we simulate the cluster.}

\item{reproduction_number}{A positive real valued scalar. The expected number
of offspring from the immigrant event.}

\item{gi_exp_decay_rate}{A positive real valued scalar. The exponential rate
of decay in the density function of the generation interval distribution
for the immigrant offspring process.}

\item{sinusoid_coefficients}{A real-valued vector of length \eqn{2 * K}.
\eqn{NULL} default value indicates no absolute-time dependance in the
offspring process.}

\item{sinusoid_frequencies}{A positive real-valued vector of length \eqn{K}.
¿\eqn{NULL} default value indicates no absolute-time dependance in the
offspring process. Angular frequencies of the sinusoidal basis.}

\item{dominating_scalar}{A positive real valued scalar greater than the
absolute-time modulating function for all \eqn{t}. Allows Lewis thinning
for simulating offspring processes. Suitable values are 2 when there is
absolute-time modulation and 2 otherwise.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
An ordered vector of real-valued event times on the sampling interval
  for the offspring of t_seed.
}
\description{
Simulate an offspring process for the event at `t_seed`.
}
