% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_clusters.R
\name{simulate_tide_hawkes_cluster}
\alias{simulate_tide_hawkes_cluster}
\title{Simulate Time-Dependant Hawkes process cluster}
\usage{
simulate_tide_hawkes_cluster(
  observed_cluster_df,
  reproduction_number,
  gi_exp_decay_rate,
  sinusoid_coefficients,
  sinusoid_frequencies,
  observation_horizon,
  simulation_horizon,
  dominating_scalar = 2,
  cluster_id = 0,
  perform_checks = TRUE
)
}
\arguments{
\item{observed_cluster_df}{A data frame containing the branching structure
and timestamps for events in the observation window. The observed events
seeding the cluster..}

\item{reproduction_number}{A positive real valued scalar. The expected number
of offspring from each event.}

\item{gi_exp_decay_rate}{A positive real valued scalar. The exponential rate
of decay in the density function of the generation interval distribution,}

\item{sinusoid_coefficients}{A real-valued vector of length \eqn{2 * K}.
Denoted \eqn{\beta} above.}

\item{sinusoid_frequencies}{A positive real-valued vector of length \eqn{K}.
Angular frequencies of the sinusoidal basis. Denoted \eqn{\omega} above.}

\item{observation_horizon}{A real valued scalar. The time up to which events
in the cluster are observed.}

\item{simulation_horizon}{A real valued scalar. The time up to which events
in the cluster are simulated.}

\item{dominating_scalar}{A positive real valued scalar greater than \eqn{\rho
(t)} for all \eqn{t}. Allows Lewis thinning for simulating the offspring
process.}

\item{cluster_id}{A scalar.An arbitrary cluster label.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A data frame with 4 variables: timestamps \eqn{t}, labels \eqn{id};
  parent labels \eqn{parent_id}; cluster label \eqn{cluster_id}.
}
\description{
Simulate a cluster generated by the time-dependant Hawkes process model up to
a specified horizon given the cluster over a specified observation window. In
this case, the offspring intensity is defined as \deqn{R \psi \rho (t) \exp
(- \psi ( t - t_i))} for \eqn{t > t_i} where \eqn{R} is the reproduction
number, \eqn{\rho (t)} is the relative activity function, and \eqn{\psi} is
the exponential decay rate for the generation interval density.
}
