% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinusoid_functions.R
\name{sinusoidal_exponential_decay_integral}
\alias{sinusoidal_exponential_decay_integral}
\title{Integrate a Linear Combination of Exponentially decaying Sinusoidal Basis
Functions}
\usage{
sinusoidal_exponential_decay_integral(
  lower_limit,
  upper_limit,
  sinusoid_coefficients,
  sinusoid_frequencies,
  exponential_rate,
  perform_checks = TRUE
)
}
\arguments{
\item{lower_limit}{A real-valued vector or scalar. The lower limit of the
integral. The quantity a above.}

\item{upper_limit}{A real-valued vector or scalar. The upper limit of the
integral. The quantity b above.}

\item{sinusoid_coefficients}{A real-valued vector of length \eqn{2 * K}.
Denoted \eqn{\beta} above.}

\item{sinusoid_frequencies}{A positive real-valued vector of length \eqn{K}.
Angular frequencies of the sinusoidal basis. Denoted \eqn{\omega} above.}

\item{exponential_rate}{A positive real-valued scalar. The exponential rate
of decay. \eqn{\psi} in the expression above.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A vector of real-valued function evaluations, one for each element of
  the longer of `lower_limit` or `upper_limit`.
}
\description{
Compute the integral of the function \deqn{\int_{a}^{b} \rho (t) \exp(- \psi
(t - a)) dt = \int_{a}^{b} \exp(- \psi (t - a)) \left( \sum_{k = 1}^K
\beta_{2k - 1} sin(\omega_k t ) + \beta_{2k} cos(\omega_k t ) \right) dt}.
}
