% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinusoid_functions.R
\name{sinusoidal_function}
\alias{sinusoidal_function}
\title{Linear Combination of Sinusoidal Basis Functions}
\usage{
sinusoidal_function(
  t,
  sinusoid_coefficients,
  sinusoid_frequencies,
  perform_checks = TRUE
)
}
\arguments{
\item{t}{A real-valued vector or scalar. The time points at which the
function is computed.}

\item{sinusoid_coefficients}{A real-valued vector of length \eqn{2 * K}.
Denoted \eqn{\beta} above.}

\item{sinusoid_frequencies}{A positive real-valued vector of length \eqn{K}.
Angular frequencies of the sinusoidal basis. Denoted \eqn{\omega} above.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A vector of real-valued function evaluations, one for each element of
  `t`.
}
\description{
Compute the value of the function \deqn{\rho (t) = \sum_{k = 1}^K \beta_{2k -
1} sin(\omega_k t ) + \beta_{2k} cos(\omega_k t )}.
}
