% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinusoid_functions.R
\name{sinusoidal_integral}
\alias{sinusoidal_integral}
\title{Integrate a Linear Combination of Sinusoidal Basis Functions}
\usage{
sinusoidal_integral(
  lower_limit,
  upper_limit,
  sinusoid_coefficients,
  sinusoid_frequencies,
  perform_checks = TRUE
)
}
\arguments{
\item{lower_limit}{A real-valued vector or scalar. The lower limit of the
integral. The quantity a above.}

\item{upper_limit}{A real-valued vector or scalar. The upper limit of the
integral. The quantity b above.}

\item{sinusoid_coefficients}{A real-valued vector of length \eqn{2 * K}.
Denoted \eqn{\beta} above.}

\item{sinusoid_frequencies}{A positive real-valued vector of length \eqn{K}.
Angular frequencies of the sinusoidal basis. Denoted \eqn{\omega} above.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A vector of real-valued function evaluations, one for each element of
  the longer of `lower_limit` or `upper_limit`.
}
\description{
Compute the integral of the function \deqn{\int_{a}^{b} \rho (t) dt =
\int_{a}^{b} \sum_{k = 1}^K \beta_{2k - 1} sin(\omega_k t ) + \beta_{2k}
cos(\omega_k t ) dt}.
}
