% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{Testing data models od online message-board activity}
\format{

}
\usage{
test_df
}
\description{
A data set containing 10% of the discussions on the \eqn{r/ireland} subreddit
seeded  by posts between April 29 and May 10, 2019 inclusive and observed for 48 hours after the
initial post. That is, we have a 3,716 discussions sampled at random after excluding discussions
already included in `train_df`.
}
\keyword{datasets}
