% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_likelihood.R
\name{tide_mixture_bpp_likelihood}
\alias{tide_mixture_bpp_likelihood}
\title{Time-dependant Mixture Branching Process Likelihood}
\usage{
tide_mixture_bpp_likelihood(
  t,
  beta,
  a,
  mu_immigrant,
  eta_immigrant,
  mu_offspring = mu_immigrant,
  eta_offspring = eta_immigrant,
  omega = NULL,
  alpha = NULL,
  log = TRUE,
  perform_checks = TRUE
)
}
\arguments{
\item{t}{A real-valued vector of length N. The times of the point process.}

\item{beta}{An integer vector of length N. The branching structure of the
point process.}

\item{a}{A real valued scalar. The upper limit of the observation interval.}

\item{mu_immigrant}{A positive real-valued scalar. The reproduction number
for immigrant offspring processes.}

\item{eta_immigrant}{A positive real-valued scalar. The memory decay rate for
immigrant offspring processes.}

\item{mu_offspring}{A positive real-valued scalar. The reproduction number
for offspring offspring processes.}

\item{eta_offspring}{A positive real-valued scalar. The memory decay rate for
offspring offspring processes.}

\item{omega}{A positive, increasing real-valued vector of length K.
Frequencies defining the sinusoidal basis functions for exogeneous
excitation.}

\item{alpha}{A real-valued vector of length 2K. Coefficients of the
sinusoidal basis function for exogenous excitation.}

\item{log}{A logical scalar. If `TRUE` then the log-likelihood is returned.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A scalar value. The likelihood for the branching point process.
}
\description{
Computes the branching point process likelihood for the time-dependant
mixture branching process described in Meagher & Friel, whereby the offspring
processes for immigrants and offspring differ and follow a circadian rhythm,
but are otherwise homogeneous.
}
\seealso{
homogeneous_bpp_likelihood mixture_bpp_likelihood
}
