% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{messageboard_df}
\alias{messageboard_df}
\title{Timestamps and branching structure for online message-board events}
\format{
A dataframe with 117,787 observations of 4 variables:
\describe{
  \item{id}{A unique, positive integer id for each event ordered by timestamp.}
  \item{parent_id}{The id for the parent of each event. Set to 0 when the event is a post.}
  \item{tree}{The id for each distinct discussion tree seeded by a post.}
  \item{time}{The timestamp for each event.}
}
}
\usage{
messageboard_df
}
\description{
A dataset containing timestamps and branching structure for activity on the
\eqn{r/ireland} subreddit for the 42 days (6 weeks) starting on April 1,
2019
}
\keyword{datasets}
