% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_clusters.R
\name{simulate_hawkes_cluster}
\alias{simulate_hawkes_cluster}
\title{Simulate Hawkes process cluster}
\usage{
simulate_hawkes_cluster(
  t_i,
  reproduction_number,
  gi_exp_decay_rate,
  horizon,
  cluster_id = 0,
  perform_checks = TRUE
)
}
\arguments{
\item{t_i}{A real valued scalar. The seeding event time.}

\item{reproduction_number}{A positive real valued scalar. The expected number
of offspring from each event.}

\item{gi_exp_decay_rate}{A positive real valued scalar. The exponential rate
of decay in the density function of the generation interval distribution,}

\item{horizon}{A real valued scalar greater than \eqn{t_i}. The time up to
which events in the cluster are simulated.}

\item{cluster_id}{A scalar.An arbitrary cluster label.}

\item{perform_checks}{A logical scalar. Should inputs be checked before
executing code.}
}
\value{
A data frame with 4 variables: timestamps \eqn{t}, labels
  \eqn{id}; parent labels \eqn{parent_id}; cluster label
  \eqn{cluster_id}.
}
\description{
Simulate a cluster generated by the Hawkes process model up to a specified
horizon.. In this case, the offsprinf intensity is defined as \deqn{R \psi
\exp (- \psi ( t - t_i))} for \eqn{t > t_i} where \eqn{R} is the reproduction
number and \eqn{\psi} is the exponential decay rate for the generation
interval density.
}
