% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{Testing data models for online message-board activity}
\format{
A dataframe with 11,321 observations of 4 variables:
\describe{
  #'  \item{id}{A unique, positive integer id for each event ordered by
 timestamp.}
 \item{parent_id}{The id for the parent of each event. Set to 0
 when the event is a post and so does not have a parent.}
 \item{discussion}{The id for each distinct discussion
 seeded by a post.}
 \item{t}{The number of hours since 00:00:00 on April 1, 2019.} }
}
\usage{
test_df
}
\description{
A data set containing 10% of the discussions on the \eqn{r/ireland} subreddit
seeded  by posts between April 1 and May 12, 2019 inclusive and observed for 48 hours after the
initial post. That is, we have a 3,716 discussions sampled at random after excluding discussions
already included in `train_df`.
}
\keyword{datasets}
