% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{Testing data models od online message-board activity}
\format{
A dataframe with 37,2022 observations of 4 variables:
\describe{
  \item{id}{A unique, positive integer id for each event ordered by timestamp.}
  \item{parent_id}{The id for the parent of each event. Set to 0 when the event is a post.}
  \item{tree}{The id for each distinct discussion tree seeded by a post.}
  \item{time}{The timestamp for each event.}
}
}
\usage{
test_df
}
\description{
A dataset containing discussion trees on the \eqn{r/ireland} subreddit
seeded between April 29 and May 10, 2019 and observed for 48 hours after the
initial post. Initially, each discussion seeded during the period is included with
probability \eqn{p = 0.1}, but those that had already appeared in the training data are
then excluded. This provides a total of 3,601 discussions consisting of 10,864
events.
}
\keyword{datasets}
