% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{Testing data for modelling online message-board activity}
\format{
A dataframe with 37,792 observations of 4 variables:
\describe{
  \item{id}{A unique, positive integer id for each event ordered by timestamp.}
  \item{parent_id}{The id for the parent of each event. Set to 0 when the event is a post.}
  \item{tree}{The id for each distinct discussion tree seeded by a post.}
  \item{time}{The timestamp for each event.}
}
}
\usage{
test_df
}
\description{
A dataset containing discussion trees on the \eqn{r/ireland} subreddit
seeded between April 29 and May 12, 2019 and observed up until May 12, 2019.
}
\keyword{datasets}
